Hangar Abilities Rework and Records fix
Records is looking pretty clean on my end.

I tried to find a solution for abilities that would have all of them in a single game object but I couldn't make it work. What we ended up with is pretty cattywompus... But I am like 99% sure it is going to work in all cases. I renamed that object abilities (good) and put a placeholder that is as good as I can get without the goofy structure named abilities (bad)